+++ src/components/About/About.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import styles from './About.module.css';
+
+const About = () => (<div className={styles.wrap}>
+        <div className={styles.container}>
+            <h1 className={styles.title}>Обо мне</h1>
+        </div>
+    </div>)
+
+export default About;
