+++ src/components/App/App.module.css
@@ -1,30 +1,19 @@
-body {
-    margin: 0 0;
+.wrap {
+    background-color: transparent;
 }
 
-.main {
-    height: 100vh;
+.menuList {
     display: flex;
-    background: linear-gradient(180deg, rgba(185, 180, 59, 0.43) 0%, 
-    rgba(151, 118, 80, 0.74) 100%, #FA4E4E 100%);
-     justify-content: center;
+    justify-content: center;
+    margin-block-start: 1em;
+    margin-block-end: 1em;
+    padding-inline-start: 40px;
 }
 
-.container {
-    display: table;
-    margin-top: 30px;
-    min-width: 500px;
-    background: linear-gradient(180deg, rgba(240, 53, 120, 0.83) 0%,
-     rgba(229, 254, 78, 0.5) 126.88%);
-    border-radius: 5px;
-}
-
-.title {
-    margin-top: 20px;
-    color:white;
-    text-align: center;
+.link {
+    text-decoration: none;
+    margin-right: 20px;
+    text-transform: uppercase;
     font-family: sans-serif;
-    font-size: 36px;
-    line-height: 42px;
-    margin-bottom: 30px;
+    color: #ffff;
 }
