+++ src/components/CheckboxDone/CheckboxDone.module.css
@@ -3,6 +3,7 @@
 }
 
 .checkboxLabel {
+    cursor: pointer;
     padding-left: 20px;
 }
