+++ src/components/CheckboxRemove/CheckboxRemove.module.css
@@ -3,6 +3,7 @@
 }
 
 .checkboxLabel {
+    cursor: pointer;
     padding-left: 20px;
     margin-left: 10px;
 }
