+++ src/components/Contacts/Contacts.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import styles from './Contacts.module.css';
+
+const Contacts = () => (<div className={styles.wrap}>
+    <div className={styles.container}>
+        <h1 className={styles.title}>Контакты</h1>
+    </div>
+</div>)
+
+export default Contacts;
